requirements.txt
Django==3.1.5
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
drf-yasg==1.20.0
requests==2.25.1

